+++ core/documentation/source/basic/changelog.rst
@@ -2,6 +2,12 @@
 ChangeLog
 ######################
 
+**v 1.2.2** (Planned)
+
+- Installed themes config variable removed
+- Now you can just place a theme in the themes directory to install it
+
+
 **v 1.2.1** (2012-05-02)
 
 - Added a cURL extension check on the getter() function (website_helper)